% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_list_names_to_col_value}
\alias{add_list_names_to_col_value}
\title{Isert names of a list of data.frames to data.frames as column value.}
\usage{
add_list_names_to_col_value(lst, col_name)
}
\arguments{
\item{lst}{named list of data.frames.}

\item{col_name}{name of a columnof a data.frame to insert the name}
}
\value{
list of data.frames.
}
\description{
Isert names of a list of data.frames to data.frames as column value.
}
\examples{
a <- list(data.frame(A = 1:3, B = 3:5), data.frame(A = 6:9, C = 1:4))
names(a) <- c("first", "second")
add_list_names_to_col_value(a, "D")
}
