% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{adjust.dataset}
\alias{adjust.dataset}
\title{Adjust dataset}
\usage{
adjust.dataset(df, col.names, value = 0)
}
\arguments{
\item{df}{data.frame which should be modified.}

\item{col.names}{columns names which should be present in output data.frame.}

\item{value}{value which should fill missing columns.}
}
\value{
data.frame containing only specified columns in that order. Previously missing columns will be fill with the specified value.
}
\description{
Add new columns to data.frame, fill them with specified value and trim columns which are absent in col.names.
}
\details{
this function is useful in adjust features of one dataset (e.g. test set) to another one (e.g. training set). Because some models need exactly the same features in the same order in test set as in training set to make correct prediction.
}
\examples{
df <- data.frame(A=1:10, B=11:20, C=21:30)
adjust.dataset(df, c("A","F","B"))
}

