% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{amino_acid_factors}
\alias{amino_acid_factors}
\title{Create factors for amino acids in PLIF}
\usage{
amino_acid_factors(plif_names)
}
\arguments{
\item{plif_name}{a vector with PLIF names, e.g ala45.ahbdonor, etc}
}
\value{
a factor with entries sorted by amino acid number
}
\description{
Create factors for amino acids in PLIF
}
\examples{
x <- paste0(sample(c("ala", "phe", "lys"), size = 6, replace = TRUE), sample.int(500, 6), ".", sample(c("ahbdonor", "ahbacceptor"), size = 6, replace = TRUE))
amino_acid_factors(x)
}
