% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{com.cor}
\alias{com.cor}
\title{Propability of difference between two correlation coefficients.}
\usage{
com.cor(x.cor, nx, y.cor, ny)
}
\arguments{
\item{x.cor}{first correlation coefficient.}

\item{nx}{number of objects used in calculation of first correlation coefficient.}

\item{y.cor}{second correlation coefficient.}

\item{ny}{number of objects used in calculation of second correlation coefficient.}
}
\value{
propability of significance of differance between two correlation coefficients.
}
\description{
Retuns propability of significance of difference between two correlation coefficients.
}
\examples{
com.cor(0.90, 100, 0.82, 150)
}

