% Generated by roxygen2 (4.0.1): do not edit by hand
\name{combine.datasets}
\alias{combine.datasets}
\title{Combine datasets}
\usage{
combine.datasets(df1, df2, value = 0)
}
\arguments{
\item{df1}{first data.frame to combine.}

\item{df2}{second data.frame to combine.}

\item{value}{value which should fill missing columns.}
}
\value{
combined data.frame.
}
\description{
Combine two datasets, add new columns absent in each of them and fill them with specified value.
}
\details{
this function is useful to rbind two data.frames with different columns.
}
\examples{
df1 <- data.frame(A=1:10, B=11:20, C=21:30)
df2 <- data.frame(E=1:10, C=11:20, B=21:30)
combine.datasets(df1, df2)
}

