% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cor_df2_long}
\alias{cor_df2_long}
\title{Calculate correlation between two data.frames and create a long data.frame as output}
\usage{
cor_df2_long(df1, df2, id_cols1, id_cols2)
}
\arguments{
\item{df1}{first data.frame}

\item{df2}{second data.frame}

\item{id_cols1}{columns which are not features (descriptors) in the first data.frame}

\item{id_cols2}{columns which are not features (descriptors) in the second data.frame}
}
\value{
returns a long data.frame with additional columns: `id1` and `id2` (referencing to the row number of the original data.frame) and `id_cols` with appended `.x` and `.y` suffixes. The column `value` will contain correlation coefficient
}
\description{
Calculate correlation between two data.frames and create a long data.frame as output
}
\details{
all columns which are not included in `id_cols` will be used for calculation of correlation. The number of feature columns should be identical in both data.frames.
}
\examples{
d1 <- data.frame(name = LETTERS[1:4], x = rnorm(4), y = rnorm(4), z = rnorm(4))
d2 <- data.frame(name = LETTERS[5:7], x = rnorm(3), y = rnorm(3), z = rnorm(3))
df <- cor_df2_long(d1, d2, "name", "name")
}
