% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cor_df_long}
\alias{cor_df_long}
\title{Calculate correlation within a data.frame and create a long data.frame as output}
\usage{
cor_df_long(df, id_cols)
}
\arguments{
\item{df}{data.frame}

\item{id_cols}{columns which are not features (descriptors)}
}
\value{
returns a long data.frame with additional columns: `id1` and `id2` (referencing to the row number of the original data.frame) and `id_cols` with appended `.x` and `.y` suffixes. The column `value` will contain correlation coefficient
}
\description{
Calculate correlation within a data.frame and create a long data.frame as output
}
\details{
all columns which are not included in `id_cols` will be used for calculation of correlation.
}
\examples{
d <- data.frame(name = LETTERS[1:4], x = rnorm(4), y = rnorm(4), z = rnorm(4))
df <- cor_df_long(d, "name")
}
