% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_folds_mc}
\alias{create_folds_mc}
\title{Creation of folds for group-out cross-validation based on Monte Carlo method.}
\usage{
create_folds_mc(v, nfolds = "auto", max_iter = 1000,
  start_opt_param = 100, error_limit = NA, seed = 0)
}
\arguments{
\item{v}{input vector of group variables.}

\item{nfolds}{maximum number of folds. But if it is not possible to create 
the specified number of folds its number will be automatically decreased.}

\item{max_iter}{maximum number of Monte Carlo iterations.}

\item{start_opt_param}{strting parameter of Monter Carlo optimization.}

\item{error_limit}{if final error will be greater than this values results 
will be discarded.}

\item{seed}{seed for random number generator to reproduce results.}
}
\value{
vector of folds numbers for each input object.
}
\description{
Function takes input vector of variables which will be used to group 
input objects. It creates folds with approaximately equal number of objects in each fold. 
All object of the same group put in the single fold.
}
\examples{
df <- data.frame(A=runif(100), B=runif(100), C=sample(LETTERS[1:10], 100, T))
res <- create_folds_mc(df$C, 5)
}
