% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dist2longdf}
\alias{dist2longdf}
\title{Convert dist object to long data.frame}
\usage{
dist2longdf(inDist)
}
\arguments{
\item{inDist}{`dist` object}
}
\value{
`data.frame` in the long format with all pairwise values
}
\description{
Convert dist object to long data.frame
}
\details{
implemented from https://stackoverflow.com/a/23476844/895544
}
\examples{
a <- data.frame(x = rnorm(100), y = rnorm(100))
d <- dist(a)
df <- dist2longdf(d)
}
