% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand.one.level.factor}
\alias{expand.one.level.factor}
\title{Expand one-level factor}
\usage{
expand.one.level.factor(vec, var.name = "var", comb.sep = ":", fill = NA)
}
\arguments{
\item{vec}{factor or character vector..}

\item{var.name}{variable name, which will be used for generation names of newly created variables.}

\item{comb.sep}{character, which will be used as a separator in newly created variable names between \code{var.name} and \code{vec} levels.}

\item{fill}{default value to represent missing (\code{NA}) values in input vector.}
}
\value{
data.frame.
}
\description{
Creates a design matrix like \code{model.matrix} and fill unused missing (\code{NA}) values with rows of 0.
}
\examples{
vec <- c("A","B",NA,"A","C")
expand.one.level.factor(vec)
}
\seealso{
\code{\link{model.matrix}}.
}
