% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.DA.reg}
\alias{get.DA.reg}
\title{Returns statistics of regression caret models with and without 
application of applicability domain.}
\usage{
get.DA.reg(model)
}
\arguments{
\item{model}{model trained with caret package.}
}
\value{
data.frame which conatains model's parameters, values of R square 
and RMSE with and without application of applicability domain.
}
\description{
Returns statistics of regression caret models with and without 
application of applicability domain.
}
\details{
Applicability domain is estimated by fragment control approach. 
If value of a variable is outside the range of that variable for training set 
then the corresponding case is outside applicability domain and is removed 
while computing model statistics
}
\examples{
\dontrun{reg <- get.DA.reg(model)}
}
