% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getBinaryStat}
\alias{getBinaryStat}
\title{Statistic for binary classification task.}
\usage{
getBinaryStat(pred, obs)
}
\arguments{
\item{pred}{vector of predicted labels (0, 1).}

\item{obs}{vector of observed labels (0, 1).}
}
\value{
names vector contained TP, TN, FP, FN, sensitivity, specificity, balanced accuracy and accuracy.
}
\description{
Returns TP, TN, FP, FN, sensitivity, specificity, balanced accuracy and accuracy for binary classification task.
}
\examples{
set.seed(42)
obs <- sample(0:1, 10, T)
pred <- sample(0:1, 10, T)
getBinaryStat(pred, obs)
}
