% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCV}
\alias{getCV}
\title{Cross-validation result of optimal caret model}
\usage{
getCV(caret.model)
}
\arguments{
\item{caret.model}{object of classs train returned by caret train function.}
}
\value{
vector of predicted values of all folds during cross-validation.
}
\description{
Returns result of cross-validation step for the optimal model obtained by caret.
}
\examples{
cv.pred <- getCV(pls.model)
}

