% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getKappa}
\alias{getKappa}
\title{Kappa value for binary classification}
\usage{
getKappa(pred, obs, classes = c(0, 1))
}
\arguments{
\item{pred}{vector of predicted labels.}

\item{obs}{vector of observed labels.}

\item{classes}{vector of length 2 with negative and positive class labels.}
}
\value{
Kappa value.
}
\description{
Returns Kappa value
}
\examples{
set.seed(42)
obs <- sample(0:1, 10, TRUE)
pred <- sample(0:1, 10, TRUE)
getKappa(pred, obs)
}
