% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%nin\%}
\alias{\%nin\%}
\title{Returns a logical vector indicating if there is a non-match or match for its left operand. Inverse action relatively to %in% function}
\usage{
... \%nin\% NA
}
\arguments{
\item{x}{vector of values to be matched}

\item{y}{vector of values to be matched against.}
}
\value{
logical vector of TRUE (non-match) and FALSE (match).
}
\description{
Returns a logical vector indicating if there is a non-match or match for its left operand. Inverse action relatively to %in% function
}
\examples{
set.seed(42)
cv.obs <- runif(50, 0, 10)
cv.pred <- cv.obs + rnorm(50)
R2test(cv.pred, cv.obs)
}
\seealso{
\code{\link{\%in\%}}.
}

