% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{groupwise_tanimoto}
\alias{groupwise_tanimoto}
\title{Calculate Tanimoto similarity coefficient beetwen two sets of folds for cross-validation.}
\usage{
groupwise_tanimoto(g1, g2)
}
\arguments{
\item{g1, g2}{input vectors with folds numbers for each objects.}
}
\value{
mean Tanimoto between input vectors considering objects grouping in folds
}
\description{
Calculate Tanimoto similarity coefficient beetwen two sets of folds for 
cross-validation considering objects grouping in each folds
}
\examples{
set.seed(42)
g1 <- sample(1:5, 100, T)
g2 <- sample(1:5, 100, T)
groupwise_tanimoto(g1, g2)
}
