% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.load}
\alias{local.load}
\title{Load Rdata container into the local variable.}
\usage{
local.load(file.name)
}
\arguments{
\item{file.name}{name of RData file to load.}
}
\value{
object contained in RData file.
}
\description{
Load Rdata container into the local variable.
}
\examples{
a <- local.load("model.RData")
}
\seealso{
\code{\link{load}} which this function wraps.
}

