% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{named.list}
\alias{named.list}
\title{Named list}
\usage{
named.list(...)
}
\arguments{
\item{...}{comma separated list of objects.}
}
\value{
list of given objects with corresponding names.
}
\description{
Creates a named list from given objects.
}
\examples{
a <- 1:10
b <- LETTERS[1:5]
named.list(a, b)
}
\seealso{
\code{\link{list}}.
}
