% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pareto_front}
\alias{pareto_front}
\title{Identify Pareto front for two dimentional problem.}
\usage{
pareto_front(x, y)
}
\arguments{
\item{x}{vector of values}

\item{y}{vector of values}
}
\value{
logical vector where TRUE correspond to point on a Pareto front
}
\description{
Identify Pareto front for two dimentional problem.
}
\details{
optimimum value is minimum of x and y by default, 
to change direction of prioritization one should change the sign of that values
}
\examples{
a <- data.frame(x = rnorm(100), y = rnorm(100))
a$front_1 <- pareto_front(a$x, a$y)
# to search for maximum values of x and minimum of y
a$front_1 <- pareto_front(-a$x, a$y)
}
