% Generated by roxygen2 (4.0.1): do not edit by hand
\name{R2test}
\alias{R2test}
\title{Determination coefficient for external test set or cross-validation.}
\usage{
R2test(ts.pred, ts.obs, ws.obs.mean = mean(ts.obs))
}
\arguments{
\item{ts.pred}{predicted values for external test set or cross-validation folds.}

\item{ts.obs}{observed values for external test set or cross-validation folds.}

\item{ws.obs.mean}{mean value for training set. If determination coefficient for cross-validation folds is calculated, then this parameter can be left as is.}
}
\value{
value of determination coefficient calculated according to formula 1-PRESS/SS.
}
\description{
Returns determination coefficient for external test set or cross-validation.
}
\examples{
set.seed(42)
cv.obs <- runif(50, 0, 10)
cv.pred <- cv.obs + rnorm(50)
R2test(cv.pred, cv.obs)
}

