% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{read.mda.dat}
\alias{read.mda.dat}
\title{Read dat-file of MDA1 program}
\usage{
read.mda.dat(fname.dat)
}
\arguments{
\item{fname.dat}{name of dat-file created by MDA1 program.}
}
\value{
data.frame of descriptors of set of chemical compounds.
}
\description{
Read dat-file contained descriptors of set of chemical compounds.
}
\examples{
df <- read.mda.dat("all_simplexes.dat")
}

