% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmse}
\alias{rmse}
\title{Calculate RMSE.}
\usage{
rmse(obs, pred)
}
\arguments{
\item{obs}{observed values.}

\item{pred}{predicted values.}
}
\value{
value of RMSE.
}
\description{
Returns root mean squared deviation.
}
\examples{
obs <- runif(100)
pred <- runif(100)
rmse(obs, pred)
}

