% Generated by roxygen2 (4.0.1): do not edit by hand
\name{select_folds}
\alias{select_folds}
\title{Select dissimilar sets of folds.}
\usage{
select_folds(folds_list, max_sets = 10, max_sim = 0.8,
  print_sim_matrix = FALSE)
}
\arguments{
\item{folds_list}{list of sets of folds, each element is a vector of folds numbers.}

\item{max_sets}{maximum number of sets which will be remained}

\item{max_sim}{maximum similarity value, only sets which have less or equal similarity to all other sets will be remained.}
}
\value{
list of sets of folds which are dissimilar to each other
}
\description{
Select most dissimilar folds sets from the input list based on groupwise Tanimoto measure.
}
\examples{
folds_list <- lapply(1:10, function(i) sample.int(5, 100, T))
selected_folds <- select_folds(folds_list, 5, 0.7)
}

