% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{sirms.read}
\alias{sirms.read}
\title{Read tab-delimited text file with sirms descriptors (output of sirms.py)}
\usage{
sirms.read(fname, ...)
}
\arguments{
\item{fname}{file name of tab-delimited text file with sirms descriptors.}

\item{...}{further arguments for \code{data.table::fread} function.}
}
\value{
data.frame.
}
\description{
Load tab-delimited text file with sirms descriptors in \code{data.frame}.
}

