% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sirms.read.svm}
\alias{sirms.read.svm}
\title{Read sirms descriptor saved in svm format.}
\usage{
sirms.read.svm(file_name)
}
\arguments{
\item{file_name}{path to filename to load/}
}
\value{
data.frame.
}
\description{
Read sirms descriptor saved in svm format.
}
\details{
there should be files with colnames and rownames in the same dir 
along with target file_name. They have to have the same name and extensions 
.colnames and .rownames.
}
