% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{spci.add_consensus}
\alias{spci.add_consensus}
\title{Add consensus (average) predictions to the data.frame with fragments contributions.}
\usage{
spci.add_consensus(df, model_names)
}
\arguments{
\item{df}{input data.frame.}

\item{model_names}{vector of model names.}
}
\value{
data.frame.
}
\description{
Add consensus (average) predictions to the data.frame with fragments contributions.
}
\details{
Model names have to correspond to column names of input data.frame.
These column names are usually consist of two parts model name of atomic property name
(svm_charge or rf_overall).
}
\examples{
df <- spci.add_consensus(df, c("svm", "rf", "gbm", "pls"))
}

