% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spci.filter_data}
\alias{spci.filter_data}
\title{Filter out fragments whose occurence is below specified limit and keep 
only specified models.}
\usage{
spci.filter_data(df, minM, minN, model_names)
}
\arguments{
\item{df}{input data.frame.}

\item{minM}{number of the lowest possible number of molecules comprising each 
fragment to keep them.}

\item{minN}{number of the lowest possible number of occurencies of fragmnets 
across the whole dataset.}

\item{model_names}{vector with model names to use for filtering}
}
\value{
data.frame.
}
\description{
Filter out fragments whose occurence is below specified limit and keep 
only specified models.
}
\details{
Model names have to correspond to column names of input data.frame. 
These column names are usually consist of two parts model name of atomic property name 
(svm_charge or rf_overall).
First five columns should contain full_names, M, N, MolId and FragID and they 
will be always kept.
}
\examples{
df <- spci.filter_data(df, 10, 10, c("svm", "rf", "gbm", "pls", "consensus"))
}
