% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spci.load_data}
\alias{spci.load_data}
\title{Read data from text file with fragments contributions.}
\usage{
spci.load_data(file_name, addN = TRUE, addM = TRUE, sep = "###")
}
\arguments{
\item{file_name}{name of input text file with fragments contributions.}

\item{addN}{logical whether or not to add number of occurencies of each fragments 
across the dataset.}

\item{addM}{logical whether or not to add number of molecules comprising 
each type of fragment.}

\item{sep}{separator between molecule and fragment name}
}
\value{
data.frame.
}
\description{
Read data from text file with fragments contributions.
}
\details{
Old version of input file had '#' as a separator.
}
\examples{
df <- spci.load_data(file_name)
}
