% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{spci.melt_data}
\alias{spci.melt_data}
\title{Melt data.frames with fragments for graphical representation .}
\usage{
spci.melt_data(df)
}
\arguments{
\item{df}{input data.frame.}
}
\value{
melted data.frame.
}
\description{
Melt data.frames with fragments for graphical representation .
}
\details{
First five columns should contain full_names, M, N, MolId and FragID.
}
\examples{
df <- spci.melt_data(df, "consensus_overall")
}

