% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{spci.reorder_data}
\alias{spci.reorder_data}
\title{Reorder factor levels representing fragment names according to specified
value and metric.}
\usage{
spci.reorder_data(df, order_by_var, FUN = median)
}
\arguments{
\item{df}{input data.frame.}

\item{order_by_var}{name of the column which will be used for reordering}

\item{FUN}{function used for summarization}
}
\value{
data.frame.
}
\description{
Reorder factor levels representing fragment names according to specified
value and metric.
}
\details{
all values of the specified column will be grouped by fragment names and
summarized with specified function which output will used for ordering (descending).
}
\examples{
df <- spci.reorder_data(df, "consensus_overall")
}

