% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tsne_for_dupl}
\alias{tsne_for_dupl}
\title{Calculate t-SME coordinates for a data.frame containing duplicated rows}
\usage{
tsne_for_dupl(df, feature_cols = NULL, ...)
}
\arguments{
\item{df}{an input data.frame}

\item{feature_cols}{columns with features to be used for creation of a map, if `NULL` - all columns will be used}

\item{...}{parameters passed to Rtsne function}
}
\value{
a data.frame containing all columns not included in `feature_cols` if any and two additional columns `x` and `y` with coordinates
}
\description{
Calculate t-SME coordinates for a data.frame containing duplicated rows
}
