% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write.as.binary}
\alias{write.as.binary}
\title{Save text data as binary to save used encoding.}
\usage{
write.as.binary(df, file.name, sep = "\\t", row.names = TRUE,
  col.names = TRUE)
}
\arguments{
\item{df}{is `data.frame` or `matrix` to save.}

\item{sep}{separator of fields.}

\item{row.names}{write row names.}

\item{col.names}{write column names.}
}
\value{
nothing.
}
\description{
Save text data as binary to save used encoding.
}
\details{
This function was created as a workaround for Windows users to properly 
save text files with UTF-8 encoding.
}
