% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_caret_folds}
\alias{create_caret_folds}
\title{Create folds of test sets to use in caret.}
\usage{
create_caret_folds(folds_list)
}
\arguments{
\item{folds_list}{list of sets of folds, each element is a vector of folds numbers.}
}
\value{
list of folds which suit the requirements of caret train function.
}
\description{
Transform list of sets of folds to list which may be passed to caret train function.
}
\examples{
folds_list <- lapply(1:10, function(i) sample.int(5, 100, T))
caret_folds_list <- create_caret_folds(folds_list)
}

