% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Non-matching comparison}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{vector of values to be matched}

\item{y}{vector of values to be matched against.}
}
\value{
logical vector of TRUE (non-match) and FALSE (match).
}
\description{
Returns a logical vector indicating if there is a non-match or match for its left operand. Inverse action relatively to \%in\% function
}
\examples{
set.seed(42)
a <- c(1,2,15)
b <- 1:10
a \%nin\% b
!(a \%in\% b)
}
\seealso{
\code{\link{\%in\%}}.
}
