% Generated by roxygen2 (4.0.1): do not edit by hand
\name{local.load}
\alias{local.load}
\title{Load Rdata container into a local variable.}
\usage{
local.load(file.name)
}
\arguments{
\item{file.name}{name of RData file to load.}
}
\value{
object contained in RData file.
}
\description{
Load Rdata container into a local variable.
}
\examples{
a <- 1:10
save(a, file="a.RData")
b <- local.load("a.RData")
}
\seealso{
\code{\link{load}} which this function wraps.
}

